fix: ducks glitching and hovering
I fixed with an increase (subtraction) in Y. It seems that the ducks may have been clipping into walls. I've also made the original duck scene be on a collision layer 3 (I think?) whilst it has a collision mask of layer 1 so that it wont collide and glitch into eachother (I dont know which is needed to do the trick or both, but it only worked after I've also raised all the Y-coordinates by 10 w/ the help of ChatGPT) (yay I love vibe coding :D) No but seriously, it's fit to be done with ChatGPT, it's difficult and boring manual labour!